move event.timer.StudyTimer to model.StudyTimer
